backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.4
redis==5.0.1
openpyxl==3.1.2
reportlab==4.0.7
django-extensions==3.2.3
django-filter==23.3
channels==4.0.0
channels-redis==4.1.0